Move commands to file
