+++ backend/cli/__init__.py
@@ -1,13 +1,13 @@
 from typing import List
 import subprocess
 import os
-import sys
 import io
 
 import click
 
 from cli.config import setup_django_sites, setup_django as configure_django
 from cli.decorators import setup_django, load_env
+from cli import cmds
 
 
 @click.group()
@@ -18,64 +18,6 @@ def cli():
 # TODO(chdsbd): Remove frontend/backend differentiators for linting and
 # formatting. We should differentiate based on language (Python, Typescript) or
 # maybe formatters (Prettier, Black) since Prettier formats multiple languages.
-
-
-def prettier(check: bool) -> str:
-    check_flag = "--list-different" if check else "--write"
-    glob = "frontend/**/*.{js,jsx,scss,css,ts,tsx,json}"
-    return f"$(yarn bin)/prettier '{glob}' {check_flag}"
-
-
-def mypy() -> str:
-    from pathlib import Path
-
-    files = " ".join(
-        str(p)
-        for p in Path("backend").rglob("*.py")
-        if ".ropeproject" not in str(p) and ".venv" not in str(p)
-    )
-
-    return f"mypy --config-file tox.ini {files}"
-
-
-def pylint() -> str:
-    from pathlib import Path
-
-    python_dirs = " ".join(
-        list(str(p.parent) for p in Path("backend").glob("*/__init__.py"))
-    )
-    return f"pylint --rcfile='.pylintrc' {python_dirs}"
-
-
-def black(check: bool) -> str:
-    check_flag = "--check" if check else ""
-    return f"black . {check_flag}"
-
-
-def tslint() -> str:
-    return "$(yarn bin)/tslint --project tsconfig.json --format 'codeFrame'"
-
-
-def typescript() -> str:
-    return "$(yarn bin)/tsc --noEmit"
-
-
-def flake8() -> str:
-    return "flake8 ."
-
-
-def pytest(watch: bool, args: List[str]) -> str:
-    args_str = " ".join(args)
-    if watch:
-        return f"ptw -- {args_str}"
-
-    return f"pytest {args_str}"
-
-
-def jest() -> str:
-    if os.getenv("CI"):
-        return "node scripts/test.js --env=jsdom --coverage --runInBand"
-    return "node scripts/test.js --env=jsdom"
 
 
 @cli.command(help="lint code")
@@ -89,18 +31,18 @@ def lint(ctx: click.core.Context, api: bool, web: bool) -> None:
 
     with ProcessManager() as m:
         if web or is_all:
-            m.add_process("tslint", tslint())
-            m.add_process("prettier", prettier(check=True))
-            m.add_process("typescript", typescript())
+            m.add_process("tslint", cmds.tslint())
+            m.add_process("prettier", cmds.prettier(check=True))
+            m.add_process("typescript", cmds.typescript())
 
         if api or is_all:
             ctx.invoke(missing_migrations)
-            m.add_process("mypy", mypy())
-            m.add_process("flake8", flake8())
-            m.add_process("black", black(check=True))
+            m.add_process("mypy", cmds.mypy())
+            m.add_process("flake8", cmds.flake8())
+            m.add_process("black", cmds.black(check=True))
 
             if os.getenv("CI"):
-                m.add_process("pylint", pylint())
+                m.add_process("pylint", cmds.pylint())
 
 
 @cli.command(help="format code")
@@ -114,10 +56,10 @@ def fmt(api: bool, web: bool, check: bool) -> None:
 
     with ProcessManager() as m:
         if api or is_all:
-            m.add_process("black", black(check=check))
+            m.add_process("black", cmds.black(check=check))
 
         if web or is_all:
-            m.add_process("prettier", prettier(check=check))
+            m.add_process("prettier", cmds.prettier(check=check))
 
 
 @cli.command(help="test services")
@@ -135,10 +77,12 @@ def test(api: bool, web: bool, watch: bool, test_args: List[str]) -> None:
 
     with ProcessManager() as m:
         if api or is_all:
-            m.add_process("api", pytest(watch=watch, args=test_args), cwd="backend")
+            m.add_process(
+                "api", cmds.pytest(watch=watch, args=test_args), cwd="backend"
+            )
 
         if web or is_all:
-            m.add_process("web", jest())
+            m.add_process("web", cmds.jest())
 
 
 @cli.command(help="start dev services")
@@ -250,7 +194,7 @@ def build(api: bool, web: bool) -> None:
         call_command("collectstatic", no_input=True)
         git_sha = os.environ["GIT_SHA"]
         subprocess.run(
-            [f"sed -i s/\<%=GIT_SHA=%\>/{git_sha}/ backend/settings.py"], check=True
+            [rf"sed -i s/\<%=GIT_SHA=%\>/{git_sha}/ backend/settings.py"], check=True
         )
         subprocess.run(["grep GIT_SHA backend/settings.py"], check=True)
